requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.5
python-multipart==0.0.20
aiosqlite==0.20.0
mistune==3.1.2
pyyaml==6.0.2
aiofiles==24.1.0
